pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot_resume"
version = "0.1.0"
description = "Streamlit app for resume-based chatbot using LangChain and Qdrant"
authors = [{ name = "Hadi", email = "hadi.supra@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "streamlit>=1.28,<2.0",
    "pandas>=1.5,<2.1",
    "tiktoken>=0.5,<0.6",
    "langchain_text_splitters>=0.0.1,<0.1",
    "langchain_community>=0.0.1,<0.1",
    "qdrant-client>=1.6,<2.0",
    "python-dotenv>=1.0,<1.1"
]

[project.optional-dependencies]
dev = [
    "black>=23.0,<24.0",
    "isort>=5.12,<6.0",
    "mypy>=1.6,<2.0"
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
